Refactor requests compatibility for Ollama provider
